[Android] Does not create recorder until required.
The code seems to assume that if the `recorder` field is set, then the
recorder has been activated and some cleanup code will be called in the
`destroy` method. This only creates the recorder object when the first
call to start recording arrives, and therefore only cleans it up if it
has been used.